Agg roles y refactorizando el codigo

+++ src/controllers/music.controller.ts
@@ -4,22 +4,26 @@ import { Request, Response, request } from "express";
 import { CreateMusicDto } from "../dtos/music/create-music.dto";
 import { UpdateMusicDto } from "../dtos/music/update-music.dto";
 import musicService from "../services/music.service";
+import userController from "./user.controller";
+import {tokenG} from '../middlewares/authJwt'
 
-module.exports = {
+class MusicController {
 
     async getList(req: Request, res: Response): Promise<Response>{
         const music = await musicService.getList()
         return res.json(music);
     }
-    ,
+    
     async getOne(req: Request, res: Response): Promise<Response>{
         const { id } = req.params;
         const music = await musicService.getOne(+id);
         return res.json(music);
     }
-    ,
+    
     async create(req: Request, res: Response): Promise<Response>{
 
+        if (tokenG.role !== 'admin') return res.status(401).send({ code: 401, message: 'You dont have permission to create!' });
+        
         const payload = req.body;
         let createMusicDto = plainToClass(CreateMusicDto, payload);
 
@@ -36,9 +40,11 @@ module.exports = {
         return res.json(await musicService.create(createMusicDto));
 
     }
-    ,
+    
     //UPDATE
     async update(req: Request, res: Response): Promise<Response>{
+
+        if (tokenG.role !== 'admin') return res.status(401).send({ code: 401, message: 'You dont have permission to update!' });
 
         const { id } = req.params;
         const payload = req.body;
@@ -61,14 +67,18 @@ module.exports = {
         
         return res.json(music); 
     }
-    ,
+    
     //DELETE
     async delete(req: Request, res: Response): Promise<Response>{
+
+        if (tokenG.role !== 'admin') return res.status(401).send({ code: 401, message: 'You dont have permission to delete!' });
 
         const {id} = req.params
         
         await musicService.delete(+id)
 
-        return res.status(204).json();
+        return res.status(202).json({ msg: 'Deleted correctly!!' });
     }
 }
+
+export default new MusicController()
